added callbacks to the alerts
