+++ swift-base/AppDelegate.swift
@@ -12,7 +12,7 @@ import IQKeyboardManagerSwift
 
 @UIApplicationMain
 class AppDelegate: UIResponder, UIApplicationDelegate {
-  
+    
   var window: UIWindow?
